src/data_generator/requirements.txt
locustio==0.11.0
ai_acc_quality
